move place method from compass to automataBlockHandPlugin
